+++ chapter-5/angry-boss.rb
@@ -1,0 +1,14 @@
+puts ""
+puts "# = = = = = = = = = = "
+puts "# Chapter 5"
+puts "# File: angry-boss.rb"
+puts "# = = = = = = = = = = "
+puts ""
+
+puts 'What you want?'
+question = gets
+puts ''
+
+puts 'WHADDAYA - HOW DARE YOU!'
+puts 'YOU\'RE FIRED!'
+puts ''